+++ backend/src/main/java/edu/netcracker/jobdealer/repository/SubmissionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SubmissionRepository extends JpaRepository<Submission, Integer> {
+public interface SubmissionRepository extends JpaRepository<Submission, Long> {
     List<Submission> findAllByTask(Task task);
 }
