+++ backend/src/main/java/edu/netcracker/jobdealer/repository/TestTaskRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface TestTaskRepository extends JpaRepository<Task, Integer> {
+public interface TestTaskRepository extends JpaRepository<Task, Long> {
 
     List<Task> findAllByName(String keyword);
